pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
